+++ backend/requirements.txt
@@ -1,25 +1,25 @@
 Django>=3.0,<4.0
 psycopg2-binary>=2.8
 appdirs==1.4.4
-asgiref==3.3.1
-astroid==2.5
-autopep8==1.5.5
+asgiref==3.3.4
+astroid==2.5.3
+autopep8==1.5.6
 colorama==0.4.4
 distlib==0.3.1
 filelock==3.0.12
-isort==5.7.0
-lazy-object-proxy==1.5.2
+isort==5.8.0
+lazy-object-proxy==1.6.0
 mccabe==0.6.1
 psycopg2==2.8.6
-pycodestyle==2.6.0
-pylint==2.7.0
+pycodestyle==2.7.0
+pylint==2.7.4
 pytz==2021.1
 six==1.15.0
 sqlparse==0.4.1
 toml==0.10.2
-virtualenv==20.4.2
+virtualenv==20.4.3
 wrapt==1.12.1
 django-environ==0.4.5
-djangorestframework==3.12.0
-PyJWT~=1.7.1
+djangorestframework==3.12.4
+PyJWT==2.0.1
 django-cors-headers===3.7.0
